refactor(telegram): move checkProtocolError to proto package
