+++ telegram/read_loop.go
@@ -106,9 +106,6 @@ func (c *Client) read(ctx context.Context, b *bin.Buffer) error {
 	if err := c.conn.Recv(ctx, b); err != nil {
 		return err
 	}
-	if err := c.checkProtocolError(b); err != nil {
-		return xerrors.Errorf("protocol: %w", err)
-	}
 
 	msg, err := c.cipher.DecryptDataFrom(c.authKey, atomic.LoadInt64(&c.session), b)
 	if err != nil {
@@ -157,7 +154,7 @@ func (c *Client) readLoop(ctx context.Context) {
 			continue
 		}
 
-		var protoErr *ProtocolErr
+		var protoErr *proto.ProtocolErr
 		if errors.As(err, &protoErr) && protoErr.Code == proto.CodeAuthKeyNotFound {
 			c.log.Warn("Re-generating keys (server not found key that we provided)")
 			if err := c.createAuthKey(ctx); err != nil {
